+++ Charon/FileInterface.py
@@ -1,10 +1,9 @@
 # Copyright (c) 2018 Ultimaker B.V.
 # libCharon is released under the terms of the LGPLv3 or higher.
+from typing import Any, Dict, List, IO, Callable
 
-from io import BufferedIOBase #To indicate that getStream must return a stream.
-from typing import Any, Dict, List, IO, Optional, Callable
+from Charon.OpenMode import OpenMode
 
-from .OpenMode import OpenMode
 
 ##  An interface for accessing files.
 #
