+++ Charon/filetypes/GCodeGzFile.py
@@ -1,5 +1,8 @@
-from Charon.filetypes.GCodeFile import GCodeFile
+# Copyright (c) 2018 Ultimaker B.V.
+# libCharon is released under the terms of the LGPLv3 or higher.
 import gzip
+
+from Charon.filetypes.GCodeFile import GCodeFile
 
 
 class GCodeGzFile(GCodeFile):
